Create Datatype.java
For each input variable and appropriate primitive , you must determine if the given primitives are capable of storing it. If yes, then print: 
n can be fitted in:
* dataType
If there is more than one appropriate data type, print each one on its own line and order them by size (i.e.: ).
If the number cannot be stored in one of the four mentioned primitives, print the line:
n can't be fitted anywhere.